create missing run dir
This PR creates a missing run directory which is required for travis-worker to operate properly. This fix was already documented in our operations manual [here](https://docs.travis-ci.com/user/enterprise/operations-manual/#travis-worker-on-Ubuntu-16.04-does-not-start), back then considered a one-off situation which *could* happen but wasn't expected to happen every time. Now after a couple of support requests with people running into this problem on a constant basis, I started investigating again, finding out that this needs to be fixed in the installer.